+++ audio_dl.py
@@ -28,7 +28,7 @@ Requirements:
 """
 from __future__ import annotations
 
-__version__ = "1.2.1"
+__version__ = "1.4"
 
 import argparse
 from collections.abc import Callable